+++ rgb-solver/src/solver/macros.rs
@@ -4,31 +4,30 @@
 #[cfg( not(target_arch = "x86_64"))]
 macro_rules! log {
     ( $( $t:tt )* ) => {
-        //web_sys::console::log_1(&format!( $( $t )* ).into());
+        web_sys::console::log_1(&format!( $( $t )* ).into());
     }
 }
 
 #[cfg( target_arch = "x86_64")]
 macro_rules! log {
     ( $( $t:tt )* ) => {
-       // println!( $( $t )* );
+        println!( $( $t )* );
     }
 }
 
 #[cfg( target_arch = "x86_64")]
 macro_rules! log_trace {
      ( $( $t:tt )* ) => {
-        //println!( $( $t )* );
+        println!( $( $t )* );
     }
 }
 
 #[cfg( not(target_arch = "x86_64"))]
 macro_rules! log_trace {
      ( $( $t:tt )* ) => {
-        //web_sys::console::log_1(&format!( $( $t )* ).into());
+        web_sys::console::log_1(&format!( $( $t )* ).into());
     }
 }
-
 
 //macro_rules! expr_nop { ($body:expr) => { $body } }
