+++ rgb-solver/src/solver/mod.rs
@@ -2,10 +2,12 @@
 #[macro_use]
 pub mod macros;
 pub mod struct_defs;
+pub mod universe;
+
 pub mod grid_state;
 pub mod van;
-pub mod public_func;
+pub mod func_public;
 
 pub(crate) mod utils;
-pub(crate) mod universe_impl;
+pub(crate) mod misc;
 pub(crate) mod tests;
