+++ rgb-solver/src/solver/struct_defs.rs
@@ -1,11 +1,7 @@
 use wasm_bindgen::prelude::*;
 use wasm_typescript_definition::TypescriptDefinition;
 
-use std::collections::vec_deque::VecDeque;
-
 use super::van::Van;
-use crate::solver::grid_state::{GridState, GridStateKey};
-use std::collections::HashSet;
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize, TypescriptDefinition, Default, Hash)]
 pub struct Color {
@@ -166,24 +162,6 @@ pub struct ChoiceOverride {
 }
 
 
-#[cfg_attr( not(target_arch = "x86_64"), wasm_bindgen())]
-#[derive(Default)]
-pub struct Universe {
-    pub(crate) data: GridState,
-
-    pub(crate) choice_override_list: Vec<ChoiceOverride>,
-
-    //below are used for calculating
-    pub(crate) queue: VecDeque<GridState>,
-
-    pub(crate) seen: HashSet<u64>,
-
-    pub(crate) success: Option<GridState>,
-
-    pub(crate) iter_count: usize,
-
-    pub(crate) cache_hits: usize
-}
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub(crate) enum Directions {
